% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicmeans.R
\name{dicmeans}
\alias{dicmeans}
\title{The dichotomic k-means algorithm.}
\usage{
dicmeans(x, n, iter.max = 20, algorithm = "Lloyd")
}
\arguments{
\item{x}{A matrix or data frame. Each row corresponds to the each data.}

\item{n}{Number of clusters.}

\item{iter.max}{Given to kmeans().}

\item{algorithm}{Gigen to kmeans().}
}
\value{
The clustering result.
}
\description{
The dichotomic k-means algorithm.
}
